Submanga: workaraound for total page number
I can't find total page number in submanga.
As workaround we will follow next button until the end of page. By doing
so, scanning page number will take times, because it's single threading.
And every time we stop the task it have to reset the page number and
rescan again to get the right total page, because it's possible we stop
the task while fmd still following the next button.